feat: relocate useChatBot hook to hooks directory and update import paths
