+++ components/chatbot/index.ts
@@ -5,4 +5,4 @@
  */
 
 export { default as ChatBot } from "./ChatBot";
-export { useChatBot } from "./use-chatbot";
+export { useChatBot } from "../../hooks/use-chatbot";
